requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==3.4.8
dnspython==2.4.1
elastic-transport==7.16
elasticsearch==7.17
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
numpy==1.25.2
outcome==1.2.0
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.3
passlib==1.7.4
prompt-toolkit==1.0.14
pycparser==2.21
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
pyfiglet==0.8.post1
Pygments==2.15.1
PyInquirer==1.0.3
PyJWT==1.7.1
pymongo==4.4.1
PySocks==1.7.1
pystalk==0.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
regex==2023.6.3
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
